Scene Prop Implementation
I used the Assets prepared by Nick and prettied up our main GameScene.
I Also added a new texture to our tiles to make them blend in a little better
A slight change in location to the goal flag was made, however the collider was not moved. playtested to ensure it works.